Looking over catch sample and recovery data.
I added the MS Word file with the PSC-spec-4.1

It's not as good as having a parseable text file, but
it might just work.  I used it to grab the tag status, but I think I could probably parse almost all of it for the different fields.  To that end I put a text version of it that I got by selecting everything and doing "pbpaste | tr '\r' '\n'".  That one looks imminently parseable.